feat: create lite version of agent only with wallet, vdr, and kms.
