+++ cmd/wallet-js-sdk/scripts/setup_agent_assets.sh
@@ -17,9 +17,12 @@ echo $(grep "@trustbloc-cicd/agent-sdk-web" "package.json")
 if [[ $(grep "@trustbloc-cicd/agent-sdk-web" "package.json") ]] ; then
   echo "finding asset in trustbloc-cicd/"
   cp -Rp node_modules/@trustbloc-cicd/agent-sdk-web/dist/assets/* public/agent-js-worker/assets
+  cp -Rp node_modules/@trustbloc-cicd/agent-sdk-web/dist/assets/agent-vcwallet-js-worker.wasm.gz public/agent-js-worker/assets
 else
   echo "finding asset in trustbloc/"
   cp -Rp node_modules/@trustbloc/agent-sdk-web/dist/assets/* public/agent-js-worker/assets
+  cp -Rp node_modules/@trustbloc/agent-sdk-web/dist/assets/agent-vcwallet-js-worker.wasm.gz public/agent-js-worker/assets
 fi
 
 gunzip public/agent-js-worker/assets/agent-js-worker.wasm.gz
+gunzip public/agent-js-worker/assets/agent-vcwallet-js-worker.wasm.gz
