+++ cmd/wallet-js-sdk/test/fixtures/config.ini
@@ -14,8 +14,8 @@ mediatorWSEndPoint = ws://localhost:10092
     agent-default-label = wallet-test-agent
     auto-accept = true
     log-level = debug
-    indexedDB-namespace = agent
-    storageType = indexedDB
-    didAnchorOrigin = http://testnet.orb.local
-    blocDomain = http://testnet.orb.local
+    indexed-db-namespace = agent
+    storage-type = indexedDB
+    did-anchor-origin = http://testnet.orb.local
+    bloc-domain = http://testnet.orb.local
     context-provider-url[] = http://localhost:10096/agent-startup-contexts.json
