+++ cmd/wallet-js-sdk/test/specs/universal/didcommwallet.spec.js
@@ -20,16 +20,18 @@ import {
 
 chai.use(chaiAsPromised);
 
-const WALLET_USER = "wallet-lite-user";
+const WALLET_USER = "wallet-user";
 const RELYING_PARTY = "relying-party-01";
 const keyType = "ED25519";
 const signatureType = "Ed25519VerificationKey2018";
 
 let walletAgent, rp, sampleMetadata;
 
-
 before(async function () {
-  walletAgent = await loadFrameworks({ name: WALLET_USER, onlyWallet: true });
+  walletAgent = await loadFrameworks({ name: WALLET_USER });
+  await connectToMediator(walletAgent, testConfig.mediatorEndPoint);
+  let conns = await getMediatorConnections(walletAgent);
+  expect(conns).to.not.empty;
 
   rp = new VerifierAdapter(RELYING_PARTY);
   await rp.init();
@@ -42,7 +44,7 @@ after(function () {
   rp ? rp.destroy() : "";
 });
 
-describe("Lite wallet tests", async function () {
+describe("Universal wallet tests", async function () {
   it("wallet user creates wallet profile", async function () {
     await createWalletProfile(walletAgent, WALLET_USER, {
       localKMSPassphrase: testConfig.walletUserPassphrase,
@@ -136,6 +138,84 @@ describe("Lite wallet tests", async function () {
     expect(keyPair.publicKey).to.not.empty;
   });
 
+  let routerConnections;
+  it("wallet connected to mediator for DIDComm", async function () {
+    await connectToMediator(walletAgent, testConfig.mediatorEndPoint);
+
+    let conns = await getMediatorConnections(walletAgent, { single: true });
+    expect(conns).to.not.empty;
+
+    routerConnections = [conns];
+  });
+
+  it("wallet accepts out-of-band invitation from a relyng party", async function () {
+    let invitation = await rp.createInvitation();
+    rp.acceptExchangeRequest();
+
+    let wallet = new UniversalWallet({ agent: walletAgent, user: WALLET_USER });
+    let { connectionID } = await wallet.connect(auth, invitation, {
+      routerConnections,
+    });
+    expect(connectionID).to.not.empty;
+  });
+
+  let request;
+  it("wallet sends message proposing presentation to relying party", async function () {
+    let invitation = await rp.createInvitation();
+    rp.acceptExchangeRequest();
+    rp.acceptPresentationProposal({
+      id: "22c77155-edf2-4ec5-8d44-b393b4e4fa38",
+      input_descriptors: [
+        {
+          id: "20b073bb-cede-4912-9e9d-334e5702077b",
+          schema: [
+            { uri: "https://www.w3.org/2018/credentials#VerifiableCredential" },
+          ],
+          constraints: {
+            fields: [{ path: ["$.credentialSubject.familyName"] }],
+          },
+        },
+      ],
+    });
+
+    let wallet = new UniversalWallet({ agent: walletAgent, user: WALLET_USER });
+    let { presentationRequest } = await wallet.proposePresentation(
+      auth,
+      invitation,
+      { routerConnections }
+    );
+    expect(presentationRequest).to.not.empty;
+    request = presentationRequest;
+  });
+
+  it("wallet sends present proof to relying party", async function () {
+    let prcVC = getJSONTestData("prc-vc.json");
+
+    let wallet = new UniversalWallet({ agent: walletAgent, user: WALLET_USER });
+    await wallet.presentProof(auth, request["~thread"].thid, {
+      "@context": [
+        "https://www.w3.org/2018/credentials/v1",
+        "https://identity.foundation/presentation-exchange/submission/v1",
+      ],
+      presentation_submission: {
+        definition_id: "22c77155-edf2-4ec5-8d44-b393b4e4fa38",
+        descriptor_map: [
+          {
+            format: "ldp_vp",
+            id: "20b073bb-cede-4912-9e9d-334e5702077b",
+            path: "$.verifiableCredential[0]",
+          },
+        ],
+        id: "947b7a1b-efd4-4aba-9b49-b6e1acedd3d7",
+      },
+      type: ["VerifiablePresentation", "PresentationSubmission"],
+      verifiableCredential: [prcVC],
+    });
+
+    let presentation = await rp.acceptPresentProof();
+    expect(presentation.verifiableCredential).to.not.empty;
+    expect(presentation.verifiableCredential[0].id).to.be.equal(prcVC.id);
+  });
 
   it("wallet user closes wallet", async function () {
     let wallet = new UniversalWallet({ agent: walletAgent, user: WALLET_USER });
