+++ scripts/check_lint.sh
@@ -22,5 +22,7 @@ echo "Linting agent-rest module..."
 ${DOCKER_CMD} run --rm -v $(pwd):/opt/workspace -w /opt/workspace/cmd/agent-rest golangci/golangci-lint:$GOLANGCI_LINT_VER golangci-lint run -c ../../.golangci.yml --path-prefix "cmd/agent-rest/"
 echo "Linting agent-js-worker module..."
 ${DOCKER_CMD} run --rm -e GOOS=js -e GOARCH=wasm -v $(pwd):/opt/workspace -w /opt/workspace/cmd/agent-js-worker golangci/golangci-lint:$GOLANGCI_LINT_VER golangci-lint run -c ../../.golangci.yml --path-prefix "cmd/agent-js-worker"
+echo "Linting agent-js-worker vcwallet module..."
+${DOCKER_CMD} run --rm -e GOOS=js -e GOARCH=wasm -v $(pwd):/opt/workspace -w /opt/workspace/cmd/agent-js-worker/vcwallet golangci/golangci-lint:$GOLANGCI_LINT_VER golangci-lint run -c ../../../.golangci.yml --path-prefix "cmd/agent-js-worker/vcwallet"
 echo "Linting agent-mobile module..."
 ${DOCKER_CMD} run --rm -v $(pwd):/opt/workspace -w /opt/workspace/cmd/agent-mobile golangci/golangci-lint:$GOLANGCI_LINT_VER golangci-lint run -c ../../.golangci.yml --path-prefix "cmd/agent-mobile/"
